Made changes in errors when creating a new user, for same-syntax purposes. Also added the type returned on 200OK's, and created a JwtTokenModel for the sole purpose of showing it to swagger. Lastly, fixed the api/users/{username}, as it was not working. Needed to alter the method called to include ratings, comments and realEstates, aswell as altering the AutoMapper profiler for Users.

+++ GroupAssignmentTeamBlue.API/Models/DtoModels/UserDto.cs
@@ -26,6 +26,6 @@ namespace GroupAssignmentTeamBlue.API.Models.DtoModels
         /// <summary>
         /// Average rating recieved
         /// </summary>
-        public int RatingAvrage { get; set; }
+        public double RatingAvrage { get; set; }
     }
 }
